release installers and binaries! (#20)
Hello! This PR uses
[cargo-dist](https://github.com/axodotdev/cargo-dist) to produce
platform specific binaries and some installers to help facilitate folks
using this cool new tool. The tool generates the workflow for you- so
this github actions workflow file is 100% generated and managed by the
by the tool.

Here's a [live
example](https://github.com/axodotdev/oxbuild/releases/tag/v0.1.1-prerelease.1)
of the generated artifacts if you want to take it for a spin!
Additionally we can add npm publish if you'd like :)